backend/requirements.txt
Django==4.0.4
psycopg2-binary==2.9.1
uwsgi==2.0.19.1
djangorestframework==3.13.1
python-decouple==3.4
factory_boy==3.2.0
black==22.3.0
isort==5.10.1
requests==2.27.1
gunicorn==20.1.0